split testsuite for better distribution
